+++ OUDS/Core/Components/Sources/Controls/Chip/OUDSFilterChip.swift
@@ -34,7 +34,7 @@ import SwiftUI
 ///
 /// ## Design documentation
 ///
-/// [unified-design-system.orange.com](https://unified-design-system.orange.com/472794e18/p/73c701-components)
+/// [unified-design-system.orange.com](https://unified-design-system.orange.com/472794e18/p/77fdea-chip/b/6578e5/t/14bf4bd854)
 ///
 /// ## Themes rendering
 ///