+++ OUDS/Core/Components/Sources/Controls/TextInput/Internal/LabelContainer.swift
@@ -11,6 +11,7 @@
 // Software description: A SwiftUI components library with code examples for Orange Unified Design System
 //
 
+#if !os(watchOS) && !os(tvOS)
 import OUDSTokensSemantic
 import SwiftUI
 
@@ -83,3 +84,4 @@ struct LabelContainer: View {
         }
     }
 }
+#endif