+++ OUDS/Core/Components/Sources/Controls/TextInput/Internal/TrailingActionContainer.swift
@@ -11,6 +11,7 @@
 // Software description: A SwiftUI components library with code examples for Orange Unified Design System
 //
 
+#if !os(watchOS) && !os(tvOS)
 import OUDSTokensSemantic
 import SwiftUI
 
@@ -73,3 +74,4 @@ struct TrailingActionContainer: View {
                    action: trailingAction.action)
     }
 }
+#endif