+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Update `text input` to display the label on two lines and in the middle when no placeholder and one line at top otherwise  (Orange-OpenSource/ouds-ios#1086) 
 
+### Fixed
+
+- `outlined` default value for `radio item` component
+
 ## [0.21.0](https://github.com/Orange-OpenSource/ouds-ios/compare/0.20.0...0.21.0) - 2025-11-05
 
 ### Added